--- flatflow/scheduler/__init__.py
@@ -1,3 +1,0 @@
-from flatflow.scheduler.profiler import LatencyProfiler
-
-__all__ = ["LatencyProfiler"]
