+++ flatflow/torch/profiler/__init__.py
@@ -1,0 +1,3 @@
+from flatflow.torch.profiler import ComputeProfiler
+
+__all__ = ["ComputeProfiler"]
